Lint, per bug #394
svn:revision:16771
cvs:account:mike
cvs:branch:trunk

+++ util/pixmorph.c
@@ -474,7 +474,7 @@ char **argv;
 	if (pa_width > 0) {
 	    pa_height = sb.st_size/(3*pa_width);
 	    fprintf(stderr, "width = %d, size = %ld, so height = %d\n",
-		   pa_width, sb.st_size, pa_height);
+		   pa_width, (long)sb.st_size, pa_height);
 	} else if (pa_height > 0) pa_width = sb.st_size/(3*pa_height);
 
 	if (pa_width <= 0 || pa_height <= 0) {
